tools/data_explorer/requirements.txt
flask>=3.0
pyodbc>=5.0
google-cloud-bigquery>=3.0
db-dtypes>=1.0
pandas>=2.0
plotly>=5.0
openpyxl>=3.1
cryptography>=42.0
graphfit
